Expand num009 to test more values, and add mingw32 output
This test checks that calling trig functions via the FFI gives the
same results as the Prelude versions.  But it uses an extreme value to
test: 1e20, and on Windows this gives slightly different results (for
unknown reasons).  However, using less extreme values gives reasonable
results, so I've added more values to the test to check that the
discrepancy is limited to the extreme - indeed it first goes wrong
around 1e19, values below that seem to be fine.