Fix bug causing polling instead of blocking in the non-threaded RTS (#5773)
This was a regression introduced accidentally in
6b1098511aaabd2c9503ee7be6da1944466f9cb4.  We were previously passing
a large time value to select() to simulate blocking, and this broke
due to a change from unsigned to signed arithmetic.  I've refactored
it to be less fragile now - we just pass NULL as the timeval parameter
to select(), which is the correct way to do blocking.