Add support for type-level "strings".
These are types that look like "this" and "that".
They are of kind `Symbol`, defined in module `GHC.TypeLits`.

For each type-level symbol `X`, we have a singleton type, `TSymbol X`.

The value of the singleton type can be named with the overloaded
constant `tSymbol`.  Here is an example:

tSymbol :: TSymbol "Hello"