Fix the printing of * (the kind).
Type and kinds use the same printing code, but the kind * is not
an infix operator and so it should not be wrapped in parens.

As is, 'parenSymOcc', can't get this right because it does not know
if it is looking at *, the type, or *, the kind.   This is why I
added a special case in 'ppr_tc'.