A build-system tweak for more readable build output.
This change reduces the (default) verbosity of the build system.
This makes it easier to spot warnings in the output and, also, it
makes it easier to estimate how far along are we in the build process
by just glancing at the output.

To get the traditional fully verbose output, set V=1, like this:

    make V=1