Make the interface-file deserialisation work right for promoted types (Trac #6054)
GHC currently uses the slightly-dodgy plan that when we proote
a TyCon to be a Kind constructor we leave it with the same Name.

That means that to make sense of a IfaceType we need to know wheter
it is really an IfaceType or an IfaceKind, because in the latter an
occurrence of (say) Maybe will be the *promoted* Maybe.

See Note [Checking IfaceTypes vs IfaceKinds] in TcIface