Change TH syntax to allow promoted kinds and kind polymorphism
The big change here is that Kind is no longer a distinct type,
it's just a type synonym for Type.  This reflects exactly what
happens in the HsSyn world, and avoids a great deal of duplication
between types and kinds.   But it is a breaking for (the few)
TH users who were using the constructors for Kind.

Thanks to lunaris and Richard Eisenberg for doing the work.