Change a few FFI imports to use CAPI
On Win64, ssize_t is 64 bit, but functions like read return 32 bit
ints. The CAPI wrapper means the C compiler takes care of doing all
the necessary casting.

Technically we should instead be making the types of the FFI imports
different on different platforms, but I think this will work out
simpler overall.

Before this fix, when the functions failed with -1, we thought they
were returning with 4294967295, and so didn't throw an exception.
This lead to a segfault in echo001(ghci).